+++ NEWS.md
@@ -1,5 +1,5 @@
 
-# nara 0.1.1.9019 2024-06-24
+# nara 0.1.1.9020 2024-06-26
 
 * Remove 'isocubes()'
 * Remove 'op' argument. 
@@ -37,6 +37,7 @@
     * `nr_rect()`
 * added `nr_blit_list()` and `nr_blit_grid()`.
 * added `nr_resize_nn()` and `nr_resize_bilinear()`. `nr_scale()` is now a wrapper around `nr_resize_nn()`
+* Fixed bug in polygon rendering
 
 # nara 0.1.1 2022-07-30